Cast some objects explicitly in layout. Use Queue methods instead of Collection where possible
Replace LinkedList with ArrayDeque in TableRenderer since we use only Queue methods here

+++ layout/src/main/java/com/itextpdf/layout/renderer/BlockRenderer.java
@@ -367,7 +367,7 @@ public abstract class BlockRenderer extends AbstractRenderer {
 
         if (!isPositioned()) {
             List<Point> rotatedPoints = new ArrayList<>();
-            getLayoutShiftAndRotatedPoints(rotatedPoints, (float) rotationPointX, rotationPointY);
+            getLayoutShiftAndRotatedPoints(rotatedPoints, (float) rotationPointX, (float) rotationPointY);
 
             Point clipLineBeg = new Point(layoutBox.getRight(), layoutBox.getTop());
             Point clipLineEnd = new Point(layoutBox.getRight(), layoutBox.getBottom());
